requirements.txt
Flask[async]==3.0.3
flask_socketio==5.3.6
flask_session==0.8.0
flask_cors==4.0.1
flask_mailman==1.1.1
Flask-Limiter==3.8.0

mongoengine==0.29.1
python-dotenv==1.0.1
pycryptodome==3.21.0
uuid==1.30
redis==5.1.1
Jinja2==3.1.4
openai==1.51.2